Module owo_colors::colored

Module for drop-in colored support to aid in porting code from colored to owo-colors.

Just replace:

use colored::*;

with

use owo_colors::colored::*;

Re-exports

pub use crate::OwoColorize;

Modules

control

A couple of functions to enable and disable coloring similarly to colored

Enums

Color

Available standard ANSI colors for use with OwoColorize::color or OwoColorize::on_color