owo_colors
Module combo

Macros

color_methods 🔒

implement specialized color methods for FgColorDisplay BgColorDisplay, ComboColorDisplay

impl_fmt_for_combo 🔒

Structs

ComboColorDisplay

A wrapper type which applies both a foreground and background color

ComboDynColorDisplay

Wrapper around a type which implements all the formatters the wrapped type does, with the addition of changing the foreground and background color. Is not recommended unless compile-time coloring is not an option.