Struct owo_colors::FgColorDisplay

#[repr(transparent)]
pub struct FgColorDisplay<'a, C: Color, T>(_, _);

Transparent wrapper around a type which implements all the formatters the wrapped type does, with the addition of changing the foreground color. Recommended to be constructed using OwoColorize.

Implementations

impl<'a, Fg, T> FgColorDisplay<'a, Fg, T> where
Fg: Color,

pub fn color<NewFg: DynColor>(
self,
fg: NewFg
) -> FgDynColorDisplay<'a, NewFg, T>

Set the foreground color at runtime. Only use if you do not know which color will be used at compile-time. If the color is constant, use either OwoColorize::fg or a color-specific method, such as OwoColorize::green,

use owo_colors::{OwoColorize, AnsiColors};

println!("{}", "green".color(AnsiColors::Green));

pub fn on_color<NewBg: DynColor>(
self,
bg: NewBg
) -> ComboDynColorDisplay<'a, Fg::DynEquivelant, NewBg, T>

Set the background color at runtime. Only use if you do not know what color to use at compile-time. If the color is constant, use either OwoColorize::bg or a color-specific method, such as OwoColorize::on_yellow,

use owo_colors::{OwoColorize, AnsiColors};

println!("{}", "yellow background".on_color(AnsiColors::BrightYellow));

pub fn fg<C: Color>(self) -> FgColorDisplay<'a, C, T>

Set the foreground color generically

use owo_colors::{OwoColorize, colors::*};

println!("{}", "red foreground".fg::<Red>());

pub fn bg<C: Color>(self) -> ComboColorDisplay<'a, Fg, C, T>

Set the background color generically.

use owo_colors::{OwoColorize, colors::*};

println!("{}", "black background".bg::<Black>());

pub fn black(self) -> FgColorDisplay<'a, Black, T>

Change the foreground color to black

pub fn on_black(self) -> ComboColorDisplay<'a, Fg, Black, T>

Change the background color to black

pub fn red(self) -> FgColorDisplay<'a, Red, T>

Change the foreground color to red

pub fn on_red(self) -> ComboColorDisplay<'a, Fg, Red, T>

Change the background color to red

pub fn green(self) -> FgColorDisplay<'a, Green, T>

Change the foreground color to green

pub fn on_green(self) -> ComboColorDisplay<'a, Fg, Green, T>

Change the background color to green

pub fn yellow(self) -> FgColorDisplay<'a, Yellow, T>

Change the foreground color to yellow

pub fn on_yellow(self) -> ComboColorDisplay<'a, Fg, Yellow, T>

Change the background color to yellow

pub fn blue(self) -> FgColorDisplay<'a, Blue, T>

Change the foreground color to blue

pub fn on_blue(self) -> ComboColorDisplay<'a, Fg, Blue, T>

Change the background color to blue

pub fn magenta(self) -> FgColorDisplay<'a, Magenta, T>

Change the foreground color to magenta

pub fn on_magenta(self) -> ComboColorDisplay<'a, Fg, Magenta, T>

Change the background color to magenta

pub fn purple(self) -> FgColorDisplay<'a, Magenta, T>

Change the foreground color to purple

pub fn on_purple(self) -> ComboColorDisplay<'a, Fg, Magenta, T>

Change the background color to purple

pub fn cyan(self) -> FgColorDisplay<'a, Cyan, T>

Change the foreground color to cyan

pub fn on_cyan(self) -> ComboColorDisplay<'a, Fg, Cyan, T>

Change the background color to cyan

pub fn white(self) -> FgColorDisplay<'a, White, T>

Change the foreground color to white

pub fn on_white(self) -> ComboColorDisplay<'a, Fg, White, T>

Change the background color to white

pub fn bright_black(self) -> FgColorDisplay<'a, BrightBlack, T>

Change the foreground color to bright black

pub fn on_bright_black(self) -> ComboColorDisplay<'a, Fg, BrightBlack, T>

Change the background color to bright black

pub fn bright_red(self) -> FgColorDisplay<'a, BrightRed, T>

Change the foreground color to bright red

pub fn on_bright_red(self) -> ComboColorDisplay<'a, Fg, BrightRed, T>

Change the background color to bright red

pub fn bright_green(self) -> FgColorDisplay<'a, BrightGreen, T>

Change the foreground color to bright green

pub fn on_bright_green(self) -> ComboColorDisplay<'a, Fg, BrightGreen, T>

Change the background color to bright green

pub fn bright_yellow(self) -> FgColorDisplay<'a, BrightYellow, T>

Change the foreground color to bright yellow

pub fn on_bright_yellow(self) -> ComboColorDisplay<'a, Fg, BrightYellow, T>

Change the background color to bright yellow

pub fn bright_blue(self) -> FgColorDisplay<'a, BrightBlue, T>

Change the foreground color to bright blue

pub fn on_bright_blue(self) -> ComboColorDisplay<'a, Fg, BrightBlue, T>

Change the background color to bright blue

pub fn bright_magenta(self) -> FgColorDisplay<'a, BrightMagenta, T>

Change the foreground color to bright magenta

pub fn on_bright_magenta(self) -> ComboColorDisplay<'a, Fg, BrightMagenta, T>

Change the background color to bright magenta

pub fn bright_purple(self) -> FgColorDisplay<'a, BrightMagenta, T>

Change the foreground color to bright purple

pub fn on_bright_purple(self) -> ComboColorDisplay<'a, Fg, BrightMagenta, T>

Change the background color to bright purple

pub fn bright_cyan(self) -> FgColorDisplay<'a, BrightCyan, T>

Change the foreground color to bright cyan

pub fn on_bright_cyan(self) -> ComboColorDisplay<'a, Fg, BrightCyan, T>

Change the background color to bright cyan

pub fn bright_white(self) -> FgColorDisplay<'a, BrightWhite, T>

Change the foreground color to bright white

pub fn on_bright_white(self) -> ComboColorDisplay<'a, Fg, BrightWhite, T>

Change the background color to bright white