Struct owo_colors::Style

pub struct Style { /* fields omitted */ }

Implementations

impl Style

pub fn new() -> Self

pub fn style<T>(&self, target: T) -> Styled<T>

pub fn fg<C: Color>(self) -> Self

Set the foreground color generically

use owo_colors::{OwoColorize, colors::*};

println!("{}", "red foreground".fg::<Red>());

pub fn bg<C: Color>(self) -> Self

Set the background color generically.

use owo_colors::{OwoColorize, colors::*};

println!("{}", "black background".bg::<Black>());

pub fn black(self) -> Self

Change the foreground color to black

pub fn on_black(self) -> Self

Change the foreground color to black

pub fn red(self) -> Self

Change the foreground color to red

pub fn on_red(self) -> Self

Change the foreground color to red

pub fn green(self) -> Self

Change the foreground color to green

pub fn on_green(self) -> Self

Change the foreground color to green

pub fn yellow(self) -> Self

Change the foreground color to yellow

pub fn on_yellow(self) -> Self

Change the foreground color to yellow

pub fn blue(self) -> Self

Change the foreground color to blue

pub fn on_blue(self) -> Self

Change the foreground color to blue

pub fn magenta(self) -> Self

Change the foreground color to magenta

pub fn on_magenta(self) -> Self

Change the foreground color to magenta

pub fn purple(self) -> Self

Change the foreground color to purple

pub fn on_purple(self) -> Self

Change the foreground color to purple

pub fn cyan(self) -> Self

Change the foreground color to cyan

pub fn on_cyan(self) -> Self

Change the foreground color to cyan

pub fn white(self) -> Self

Change the foreground color to white

pub fn on_white(self) -> Self

Change the foreground color to white

pub fn bright_black(self) -> Self

Change the foreground color to bright black

pub fn on_bright_black(self) -> Self

Change the foreground color to bright black

pub fn bright_red(self) -> Self

Change the foreground color to bright red

pub fn on_bright_red(self) -> Self

Change the foreground color to bright red

pub fn bright_green(self) -> Self

Change the foreground color to bright green

pub fn on_bright_green(self) -> Self

Change the foreground color to bright green

pub fn bright_yellow(self) -> Self

Change the foreground color to bright yellow

pub fn on_bright_yellow(self) -> Self

Change the foreground color to bright yellow

pub fn bright_blue(self) -> Self

Change the foreground color to bright blue

pub fn on_bright_blue(self) -> Self

Change the foreground color to bright blue

pub fn bright_magenta(self) -> Self

Change the foreground color to bright magenta

pub fn on_bright_magenta(self) -> Self

Change the foreground color to bright magenta

pub fn bright_purple(self) -> Self

Change the foreground color to bright purple

pub fn on_bright_purple(self) -> Self

Change the foreground color to bright purple

pub fn bright_cyan(self) -> Self

Change the foreground color to bright cyan

pub fn on_bright_cyan(self) -> Self

Change the foreground color to bright cyan

pub fn bright_white(self) -> Self

Change the foreground color to bright white

pub fn on_bright_white(self) -> Self

Change the foreground color to bright white

pub fn bold(self) -> Self

Make the text bold

pub fn dimmed(self) -> Self

Make the text dim

pub fn italic(self) -> Self

Make the text italicized

pub fn underline(self) -> Self

Make the text italicized

pub fn blink(self) -> Self

Make the text blink

pub fn blink_fast(self) -> Self

Make the text blink (but fast!)

pub fn reversed(self) -> Self

Swap the foreground and background colors

pub fn hidden(self) -> Self

Hide the text

pub fn strikethrough(self) -> Self

Cross out the text

pub fn effect(self, effect: Effect) -> Self

pub fn remove_effect(self, effect: Effect) -> Self

pub fn effects(self, effects: &[Effect]) -> Self

pub fn remove_effects(self, effects: &[Effect]) -> Self

pub fn remove_all_effects(self) -> Self

pub fn color<Color: DynColor>(self, color: Color) -> Self

Set the foreground color at runtime. Only use if you do not know which color will be used at compile-time. If the color is constant, use either OwoColorize::fg or a color-specific method, such as OwoColorize::green,

use owo_colors::{OwoColorize, AnsiColors};

fn main() {
    println!("{}", "green".color(AnsiColors::Green));
}

pub fn on_color<Color: DynColor>(self, color: Color) -> Self

Set the background color at runtime. Only use if you do not know what color to use at compile-time. If the color is constant, use either OwoColorize::bg or a color-specific method, such as OwoColorize::on_yellow,

use owo_colors::{OwoColorize, AnsiColors};

fn main() {
    println!("{}", "yellow background".on_color(AnsiColors::BrightYellow));
}

pub fn truecolor(self, r: u8, g: u8, b: u8) -> Self

Sets the foreground color to an RGB value.

pub fn on_truecolor(self, r: u8, g: u8, b: u8) -> Self

Sets the background color to an RGB value.