owo_colors
Module styled_list

Modules

sealed 🔒

Structs

StyledList

A collection of Styled items that are displayed in such a way as to minimize the amount of characters that are written when displayed.

Enums

Transition 🔒

How the transition between two styles should be printed