Struct owo_colors::styles::BlinkFastDisplay

#[repr(transparent)]
pub struct BlinkFastDisplay<'a, T>(pub &'a T);

Transparent wrapper around a type which implements all the formatters the wrapped type does, with the addition of making it blink fast. Recommended to be constructed using OwoColorize.

Tuple Fields
0: &'a T