owo_colors
Trait OwoColorize

pub trait OwoColorize: Sized {

    // Provided methods
    fn fg<C: Color>(&self) -> FgColorDisplay<'_, C, Self> { ... }
    fn bg<C: Color>(&self) -> BgColorDisplay<'_, C, Self> { ... }
    fn black(&self) -> FgColorDisplay<'_, Black, Self> { ... }
    fn on_black(&self) -> BgColorDisplay<'_, Black, Self> { ... }
    fn red(&self) -> FgColorDisplay<'_, Red, Self> { ... }
    fn on_red(&self) -> BgColorDisplay<'_, Red, Self> { ... }
    fn green(&self) -> FgColorDisplay<'_, Green, Self> { ... }
    fn on_green(&self) -> BgColorDisplay<'_, Green, Self> { ... }
    fn yellow(&self) -> FgColorDisplay<'_, Yellow, Self> { ... }
    fn on_yellow(&self) -> BgColorDisplay<'_, Yellow, Self> { ... }
    fn blue(&self) -> FgColorDisplay<'_, Blue, Self> { ... }
    fn on_blue(&self) -> BgColorDisplay<'_, Blue, Self> { ... }
    fn magenta(&self) -> FgColorDisplay<'_, Magenta, Self> { ... }
    fn on_magenta(&self) -> BgColorDisplay<'_, Magenta, Self> { ... }
    fn purple(&self) -> FgColorDisplay<'_, Magenta, Self> { ... }
    fn on_purple(&self) -> BgColorDisplay<'_, Magenta, Self> { ... }
    fn cyan(&self) -> FgColorDisplay<'_, Cyan, Self> { ... }
    fn on_cyan(&self) -> BgColorDisplay<'_, Cyan, Self> { ... }
    fn white(&self) -> FgColorDisplay<'_, White, Self> { ... }
    fn on_white(&self) -> BgColorDisplay<'_, White, Self> { ... }
    fn default_color(&self) -> FgColorDisplay<'_, Default, Self> { ... }
    fn on_default_color(&self) -> BgColorDisplay<'_, Default, Self> { ... }
    fn bright_black(&self) -> FgColorDisplay<'_, BrightBlack, Self> { ... }
    fn on_bright_black(&self) -> BgColorDisplay<'_, BrightBlack, Self> { ... }
    fn bright_red(&self) -> FgColorDisplay<'_, BrightRed, Self> { ... }
    fn on_bright_red(&self) -> BgColorDisplay<'_, BrightRed, Self> { ... }
    fn bright_green(&self) -> FgColorDisplay<'_, BrightGreen, Self> { ... }
    fn on_bright_green(&self) -> BgColorDisplay<'_, BrightGreen, Self> { ... }
    fn bright_yellow(&self) -> FgColorDisplay<'_, BrightYellow, Self> { ... }
    fn on_bright_yellow(&self) -> BgColorDisplay<'_, BrightYellow, Self> { ... }
    fn bright_blue(&self) -> FgColorDisplay<'_, BrightBlue, Self> { ... }
    fn on_bright_blue(&self) -> BgColorDisplay<'_, BrightBlue, Self> { ... }
    fn bright_magenta(&self) -> FgColorDisplay<'_, BrightMagenta, Self> { ... }
    fn on_bright_magenta(&self) -> BgColorDisplay<'_, BrightMagenta, Self> { ... }
    fn bright_purple(&self) -> FgColorDisplay<'_, BrightMagenta, Self> { ... }
    fn on_bright_purple(&self) -> BgColorDisplay<'_, BrightMagenta, Self> { ... }
    fn bright_cyan(&self) -> FgColorDisplay<'_, BrightCyan, Self> { ... }
    fn on_bright_cyan(&self) -> BgColorDisplay<'_, BrightCyan, Self> { ... }
    fn bright_white(&self) -> FgColorDisplay<'_, BrightWhite, Self> { ... }
    fn on_bright_white(&self) -> BgColorDisplay<'_, BrightWhite, Self> { ... }
    fn bold(&self) -> BoldDisplay<'_, Self> { ... }
    fn dimmed(&self) -> DimDisplay<'_, Self> { ... }
    fn italic(&self) -> ItalicDisplay<'_, Self> { ... }
    fn underline(&self) -> UnderlineDisplay<'_, Self> { ... }
    fn blink(&self) -> BlinkDisplay<'_, Self> { ... }
    fn blink_fast(&self) -> BlinkFastDisplay<'_, Self> { ... }
    fn reversed(&self) -> ReversedDisplay<'_, Self> { ... }
    fn hidden(&self) -> HiddenDisplay<'_, Self> { ... }
    fn strikethrough(&self) -> StrikeThroughDisplay<'_, Self> { ... }
    fn color<Color: DynColor>(
        &self,
        color: Color,
    ) -> FgDynColorDisplay<'_, Color, Self> { ... }
    fn on_color<Color: DynColor>(
        &self,
        color: Color,
    ) -> BgDynColorDisplay<'_, Color, Self> { ... }
    fn fg_rgb<const R: u8, const G: u8, const B: u8>(
        &self,
    ) -> FgColorDisplay<'_, CustomColor<R, G, B>, Self> { ... }
    fn bg_rgb<const R: u8, const G: u8, const B: u8>(
        &self,
    ) -> BgColorDisplay<'_, CustomColor<R, G, B>, Self> { ... }
    fn truecolor(&self, r: u8, g: u8, b: u8) -> FgDynColorDisplay<'_, Rgb, Self> { ... }
    fn on_truecolor(
        &self,
        r: u8,
        g: u8,
        b: u8,
    ) -> BgDynColorDisplay<'_, Rgb, Self> { ... }
    fn style(&self, style: Style) -> Styled<&Self> { ... }

}

Extension trait for colorizing a type which implements any std formatter (Display, Debug, UpperHex, etc.)

Example

use owo_colors::OwoColorize;

println!("My number is {:#x}!", 10.green());
println!("My number is not {}!", 4.on_red());

How to decide which method to use

Do you have a specific color you want to use?

Use the specific color’s method, such as blue or on_green.

Do you want your colors configurable via generics?

Use fg and bg to make it compile-time configurable.

Do you need to pick a color at runtime?

Use the color, on_color, truecolor or on_truecolor.

Do you need some other text modifier?

bold

dimmed

italic

underline

blink

blink_fast

reversed

hidden

strikethrough

Do you want it to only display colors if it’s a terminal?

Enable the supports-colors feature

Colorize inside if_supports_color

Do you need to store a set of colors/effects to apply to multiple things?

Use style to apply a Style

Provided Methods

fn fg<C: Color>(&self) -> FgColorDisplay<'_, C, Self>

Set the foreground color generically

use owo_colors::{OwoColorize, colors::*};

println!("{}", "red foreground".fg::<Red>());

fn bg<C: Color>(&self) -> BgColorDisplay<'_, C, Self>

Set the background color generically.

use owo_colors::{OwoColorize, colors::*};

println!("{}", "black background".bg::<Black>());

fn black(&self) -> FgColorDisplay<'_, Black, Self>

Change the foreground color to black

fn on_black(&self) -> BgColorDisplay<'_, Black, Self>

Change the background color to black

fn red(&self) -> FgColorDisplay<'_, Red, Self>

Change the foreground color to red

fn on_red(&self) -> BgColorDisplay<'_, Red, Self>

Change the background color to red

fn green(&self) -> FgColorDisplay<'_, Green, Self>

Change the foreground color to green

fn on_green(&self) -> BgColorDisplay<'_, Green, Self>

Change the background color to green

fn yellow(&self) -> FgColorDisplay<'_, Yellow, Self>

Change the foreground color to yellow

fn on_yellow(&self) -> BgColorDisplay<'_, Yellow, Self>

Change the background color to yellow

fn blue(&self) -> FgColorDisplay<'_, Blue, Self>

Change the foreground color to blue

fn on_blue(&self) -> BgColorDisplay<'_, Blue, Self>

Change the background color to blue

fn magenta(&self) -> FgColorDisplay<'_, Magenta, Self>

Change the foreground color to magenta

fn on_magenta(&self) -> BgColorDisplay<'_, Magenta, Self>

Change the background color to magenta

fn purple(&self) -> FgColorDisplay<'_, Magenta, Self>

Change the foreground color to purple

fn on_purple(&self) -> BgColorDisplay<'_, Magenta, Self>

Change the background color to purple

fn cyan(&self) -> FgColorDisplay<'_, Cyan, Self>

Change the foreground color to cyan

fn on_cyan(&self) -> BgColorDisplay<'_, Cyan, Self>

Change the background color to cyan

fn white(&self) -> FgColorDisplay<'_, White, Self>

Change the foreground color to white

fn on_white(&self) -> BgColorDisplay<'_, White, Self>

Change the background color to white

fn default_color(&self) -> FgColorDisplay<'_, Default, Self>

Change the foreground color to the terminal default

fn on_default_color(&self) -> BgColorDisplay<'_, Default, Self>

Change the background color to the terminal default

fn bright_black(&self) -> FgColorDisplay<'_, BrightBlack, Self>

Change the foreground color to bright black

fn on_bright_black(&self) -> BgColorDisplay<'_, BrightBlack, Self>

Change the background color to bright black

fn bright_red(&self) -> FgColorDisplay<'_, BrightRed, Self>

Change the foreground color to bright red

fn on_bright_red(&self) -> BgColorDisplay<'_, BrightRed, Self>

Change the background color to bright red

fn bright_green(&self) -> FgColorDisplay<'_, BrightGreen, Self>

Change the foreground color to bright green

fn on_bright_green(&self) -> BgColorDisplay<'_, BrightGreen, Self>

Change the background color to bright green

fn bright_yellow(&self) -> FgColorDisplay<'_, BrightYellow, Self>

Change the foreground color to bright yellow

fn on_bright_yellow(&self) -> BgColorDisplay<'_, BrightYellow, Self>

Change the background color to bright yellow

fn bright_blue(&self) -> FgColorDisplay<'_, BrightBlue, Self>

Change the foreground color to bright blue

fn on_bright_blue(&self) -> BgColorDisplay<'_, BrightBlue, Self>

Change the background color to bright blue

fn bright_magenta(&self) -> FgColorDisplay<'_, BrightMagenta, Self>

Change the foreground color to bright magenta

fn on_bright_magenta(&self) -> BgColorDisplay<'_, BrightMagenta, Self>

Change the background color to bright magenta

fn bright_purple(&self) -> FgColorDisplay<'_, BrightMagenta, Self>

Change the foreground color to bright purple

fn on_bright_purple(&self) -> BgColorDisplay<'_, BrightMagenta, Self>

Change the background color to bright purple

fn bright_cyan(&self) -> FgColorDisplay<'_, BrightCyan, Self>

Change the foreground color to bright cyan

fn on_bright_cyan(&self) -> BgColorDisplay<'_, BrightCyan, Self>

Change the background color to bright cyan

fn bright_white(&self) -> FgColorDisplay<'_, BrightWhite, Self>

Change the foreground color to bright white

fn on_bright_white(&self) -> BgColorDisplay<'_, BrightWhite, Self>

Change the background color to bright white

fn bold(&self) -> BoldDisplay<'_, Self>

Make the text bold

fn dimmed(&self) -> DimDisplay<'_, Self>

Make the text dim

fn italic(&self) -> ItalicDisplay<'_, Self>

Make the text italicized

fn underline(&self) -> UnderlineDisplay<'_, Self>

Make the text underlined

fn blink(&self) -> BlinkDisplay<'_, Self>

Make the text blink

fn blink_fast(&self) -> BlinkFastDisplay<'_, Self>

Make the text blink (but fast!)

fn reversed(&self) -> ReversedDisplay<'_, Self>

Swap the foreground and background colors

fn hidden(&self) -> HiddenDisplay<'_, Self>

Hide the text

fn strikethrough(&self) -> StrikeThroughDisplay<'_, Self>

Cross out the text

fn color<Color: DynColor>( &self, color: Color, ) -> FgDynColorDisplay<'_, Color, Self>

Set the foreground color at runtime. Only use if you do not know which color will be used at compile-time. If the color is constant, use either OwoColorize::fg or a color-specific method, such as OwoColorize::green,

use owo_colors::{OwoColorize, AnsiColors};

println!("{}", "green".color(AnsiColors::Green));

fn on_color<Color: DynColor>( &self, color: Color, ) -> BgDynColorDisplay<'_, Color, Self>

Set the background color at runtime. Only use if you do not know what color to use at compile-time. If the color is constant, use either OwoColorize::bg or a color-specific method, such as OwoColorize::on_yellow,

use owo_colors::{OwoColorize, AnsiColors};

println!("{}", "yellow background".on_color(AnsiColors::BrightYellow));

fn fg_rgb<const R: u8, const G: u8, const B: u8>( &self, ) -> FgColorDisplay<'_, CustomColor<R, G, B>, Self>

Set the foreground color to a specific RGB value.

fn bg_rgb<const R: u8, const G: u8, const B: u8>( &self, ) -> BgColorDisplay<'_, CustomColor<R, G, B>, Self>

Set the background color to a specific RGB value.

fn truecolor(&self, r: u8, g: u8, b: u8) -> FgDynColorDisplay<'_, Rgb, Self>

Sets the foreground color to an RGB value.

fn on_truecolor(&self, r: u8, g: u8, b: u8) -> BgDynColorDisplay<'_, Rgb, Self>

Sets the background color to an RGB value.

fn style(&self, style: Style) -> Styled<&Self>

Apply a runtime-determined style

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.