Struct owo_colors::Style

pub struct Style { /* private fields */ }

A pre-computed style that can be applied to a struct using OwoColorize::style.

Its interface mimics that of OwoColorize, but instead of chaining methods on your object, you instead chain them on the Style object before applying it.

use owo_colors::{OwoColorize, Style};

let my_style = Style::new()
    .red()
    .on_white()
    .strikethrough();

println!("{}", "red text, white background, struck through".style(my_style));

Implementations

impl Style

pub const fn new() -> Self

Create a new style to be applied later

pub const fn style<T>(&self, target: T) -> Styled<T>

Apply the style to a given struct to output.

Example

Usage in const contexts:

use owo_colors::{OwoColorize, Style, Styled};

const STYLED_TEXT: Styled<&'static str> = Style::new().bold().italic().style("bold and italic text");

println!("{}", STYLED_TEXT);

pub const fn fg<C: Color>(self) -> Self

Set the foreground color generically

use owo_colors::{OwoColorize, colors::*};

println!("{}", "red foreground".fg::<Red>());

pub const fn bg<C: Color>(self) -> Self

Set the background color generically.

use owo_colors::{OwoColorize, colors::*};

println!("{}", "black background".bg::<Black>());

pub const fn remove_fg(self) -> Self

Removes the foreground color from the style. Note that this does not apply the default color, but rather represents not changing the current terminal color.

If you wish to actively change the terminal color back to the default, see Style::default_color.

pub const fn remove_bg(self) -> Self

Removes the background color from the style. Note that this does not apply the default color, but rather represents not changing the current terminal color.

If you wish to actively change the terminal color back to the default, see Style::on_default_color.

pub const fn black(self) -> Self

Change the foreground color to black

pub const fn on_black(self) -> Self

Change the foreground color to black

pub const fn red(self) -> Self

Change the foreground color to red

pub const fn on_red(self) -> Self

Change the foreground color to red

pub const fn green(self) -> Self

Change the foreground color to green

pub const fn on_green(self) -> Self

Change the foreground color to green

pub const fn yellow(self) -> Self

Change the foreground color to yellow

pub const fn on_yellow(self) -> Self

Change the foreground color to yellow

pub const fn blue(self) -> Self

Change the foreground color to blue

pub const fn on_blue(self) -> Self

Change the foreground color to blue

pub const fn magenta(self) -> Self

Change the foreground color to magenta

pub const fn on_magenta(self) -> Self

Change the foreground color to magenta

pub const fn purple(self) -> Self

Change the foreground color to purple

pub const fn on_purple(self) -> Self

Change the foreground color to purple

pub const fn cyan(self) -> Self

Change the foreground color to cyan

pub const fn on_cyan(self) -> Self

Change the foreground color to cyan

pub const fn white(self) -> Self

Change the foreground color to white

pub const fn on_white(self) -> Self

Change the foreground color to white

pub const fn default_color(self) -> Self

Change the foreground color to the terminal default

pub const fn on_default_color(self) -> Self

Change the foreground color to the terminal default

pub const fn bright_black(self) -> Self

Change the foreground color to bright black

pub const fn on_bright_black(self) -> Self

Change the foreground color to bright black

pub const fn bright_red(self) -> Self

Change the foreground color to bright red

pub const fn on_bright_red(self) -> Self

Change the foreground color to bright red

pub const fn bright_green(self) -> Self

Change the foreground color to bright green

pub const fn on_bright_green(self) -> Self

Change the foreground color to bright green

pub const fn bright_yellow(self) -> Self

Change the foreground color to bright yellow

pub const fn on_bright_yellow(self) -> Self

Change the foreground color to bright yellow

pub const fn bright_blue(self) -> Self

Change the foreground color to bright blue

pub const fn on_bright_blue(self) -> Self

Change the foreground color to bright blue

pub const fn bright_magenta(self) -> Self

Change the foreground color to bright magenta

pub const fn on_bright_magenta(self) -> Self

Change the foreground color to bright magenta

pub const fn bright_purple(self) -> Self

Change the foreground color to bright purple

pub const fn on_bright_purple(self) -> Self

Change the foreground color to bright purple

pub const fn bright_cyan(self) -> Self

Change the foreground color to bright cyan

pub const fn on_bright_cyan(self) -> Self

Change the foreground color to bright cyan

pub const fn bright_white(self) -> Self

Change the foreground color to bright white

pub const fn on_bright_white(self) -> Self

Change the foreground color to bright white

pub const fn bold(self) -> Self

Make the text bold

pub const fn dimmed(self) -> Self

Make the text dim

pub const fn italic(self) -> Self

Make the text italicized

pub const fn underline(self) -> Self

Make the text underlined

pub const fn blink(self) -> Self

Make the text blink

pub const fn blink_fast(self) -> Self

Make the text blink (but fast!)

pub const fn reversed(self) -> Self

Swap the foreground and background colors

pub const fn hidden(self) -> Self

Hide the text

pub const fn strikethrough(self) -> Self

Cross out the text

pub const fn effect(self, effect: Effect) -> Self

Apply a given effect from the style

pub const fn remove_effect(self, effect: Effect) -> Self

Remove a given effect from the style

pub const fn effects(self, effects: &[Effect]) -> Self

Apply a given set of effects to the style

pub const fn remove_effects(self, effects: &[Effect]) -> Self

Remove a given set of effects from the style

pub const fn remove_all_effects(self) -> Self

Disables all the given effects from the style

pub fn color<Color: DynColor>(self, color: Color) -> Self

Set the foreground color at runtime. Only use if you do not know which color will be used at compile-time. If the color is constant, use either OwoColorize::fg or a color-specific method, such as OwoColorize::green,

use owo_colors::{OwoColorize, AnsiColors};

println!("{}", "green".color(AnsiColors::Green));

pub fn on_color<Color: DynColor>(self, color: Color) -> Self

Set the background color at runtime. Only use if you do not know what color to use at compile-time. If the color is constant, use either OwoColorize::bg or a color-specific method, such as OwoColorize::on_yellow,

use owo_colors::{OwoColorize, AnsiColors};

println!("{}", "yellow background".on_color(AnsiColors::BrightYellow));

pub const fn fg_rgb<const R: u8, const G: u8, const B: u8>(self) -> Self

Set the foreground color to a specific RGB value.

pub const fn bg_rgb<const R: u8, const G: u8, const B: u8>(self) -> Self

Set the background color to a specific RGB value.

pub const fn truecolor(self, r: u8, g: u8, b: u8) -> Self

Sets the foreground color to an RGB value.

pub const fn on_truecolor(self, r: u8, g: u8, b: u8) -> Self

Sets the background color to an RGB value.

pub const fn is_plain(&self) -> bool

Returns true if the style does not apply any formatting.

pub const fn prefix_formatter(&self) -> StylePrefixFormatter

Returns a formatter for the style’s ANSI prefix.

This can be used to separate out the prefix and suffix of a style.

Example

use owo_colors::Style;
use std::fmt::Write;

let style = Style::new().red().on_blue();
let prefix = style.prefix_formatter();
let suffix = style.suffix_formatter();

// Write the prefix and suffix separately.
let mut output = String::new();
write!(output, "{}", prefix);
output.push_str("Hello");
write!(output, "{}", suffix);

assert_eq!(output, "\x1b[31;44mHello\x1b[0m");

pub const fn suffix_formatter(&self) -> StyleSuffixFormatter

Returns a formatter for the style’s ANSI suffix.

This can be used to separate out the prefix and suffix of a style.

Example

See Style::prefix_formatter.

pub fn fmt_prefix(&self, f: &mut Formatter<'_>) -> Result

Applies the ANSI-prefix for this style to the given formatter

pub fn fmt_suffix(&self, f: &mut Formatter<'_>) -> Result

Applies the ANSI-suffix for this style to the given formatter