owo_colors
Module styles

Different display styles (strikethrough, bold, etc.)

Structs

BlinkDisplay

Transparent wrapper around a type which implements all the formatters the wrapped type does, with the addition of making it blink. Recommended to be constructed using OwoColorize.

BlinkFastDisplay

Transparent wrapper around a type which implements all the formatters the wrapped type does, with the addition of making it blink fast. Recommended to be constructed using OwoColorize.

BoldDisplay

Transparent wrapper around a type which implements all the formatters the wrapped type does, with the addition of boldening it. Recommended to be constructed using OwoColorize.

DimDisplay

Transparent wrapper around a type which implements all the formatters the wrapped type does, with the addition of dimming it. Recommended to be constructed using OwoColorize.

HiddenDisplay

Transparent wrapper around a type which implements all the formatters the wrapped type does, with the addition of hiding the text. Recommended to be constructed using OwoColorize.

ItalicDisplay

Transparent wrapper around a type which implements all the formatters the wrapped type does, with the addition of italicizing it. Recommended to be constructed using OwoColorize.

ReversedDisplay

Transparent wrapper around a type which implements all the formatters the wrapped type does, with the addition of swapping foreground and background colors. Recommended to be constructed using OwoColorize.

StrikeThroughDisplay

Transparent wrapper around a type which implements all the formatters the wrapped type does, with the addition of crossing out the given text. Recommended to be constructed using OwoColorize.

UnderlineDisplay

Transparent wrapper around a type which implements all the formatters the wrapped type does, with the addition of underlining it. Recommended to be constructed using OwoColorize.