owo_colors

Module styles

Different display styles (strikethrough, bold, etc.)

Structs

BlinkDisplay

Transparent wrapper around a type which implements all the formatters the wrapped type does, while blinking. Recommended to be constructed using OwoColorize.

BlinkFastDisplay

Transparent wrapper around a type which implements all the formatters the wrapped type does, with the addition of making it blink fast. Use OwoColorize

BoldDisplay

Transparent wrapper around a type which implements all the formatters the wrapped type does, with the addition of boldening it. Recommended to be constructed using OwoColorize.

DimDisplay

Transparent wrapper around a type which implements all the formatters the wrapped type does, with the addition of dimming it. Recommended to be constructed using OwoColorize.

HiddenDisplay

Transparent wrapper around a type which implements all the formatters the wrapped type does, with the addition of hiding the text. Use OwoColorize.

ItalicDisplay

Transparent wrapper around a type which implements all the formatters the wrapped type does, with the addition of italics. Recommended to be constructed using OwoColorize.

ReversedDisplay

Transparent wrapper around a type which implements all the formatters the wrapped type does, with the addition of swapping fg and bg colors. Use OwoColorize

StrikeThroughDisplay

Transparent wrapper around a type which implements all the formatters the wrapped type does, crossed out. Recommended to be constructed using OwoColorize.

UnderlineDisplay

Transparent wrapper around a type which implements all the formatters the wrapped type does, while underlining it. Recommended to be constructed using OwoColorize.