Trait owo_colors::Color

pub trait Color: Sealed {
    const ANSI_FG: &'static str;
    const ANSI_BG: &'static str;
    const RAW_ANSI_FG: &'static str;
    const RAW_ANSI_BG: &'static str;
}

A trait for describing a type which can be used with FgColorDisplay or BgColorDisplay

Required Associated Constants

const ANSI_FG: &'static str

The ANSI format code for setting this color as the foreground

const ANSI_BG: &'static str

The ANSI format code for setting this color as the background

const RAW_ANSI_FG: &'static str

The raw ANSI format for settings this color as the foreground without the ANSI delimiters (“\x1b” and “m”)

const RAW_ANSI_BG: &'static str

The raw ANSI format for settings this color as the background without the ANSI delimiters (“\x1b” and “m”)

Object Safety

This trait is not object safe.