Trait owo_colors::DynColor

pub trait DynColor: Sealed {
    // Required methods
    fn fmt_ansi_fg(&self, f: &mut Formatter<'_>) -> Result;
    fn fmt_ansi_bg(&self, f: &mut Formatter<'_>) -> Result;
    fn fmt_raw_ansi_fg(&self, f: &mut Formatter<'_>) -> Result;
    fn fmt_raw_ansi_bg(&self, f: &mut Formatter<'_>) -> Result;
}

A trait describing a runtime-configurable color which can displayed using FgDynColorDisplay or BgDynColorDisplay. If your color will be known at compile time it is recommended you avoid this.

Required Methods

fn fmt_ansi_fg(&self, f: &mut Formatter<'_>) -> Result

A function to output a ANSI code to a formatter to set the foreground to this color

fn fmt_ansi_bg(&self, f: &mut Formatter<'_>) -> Result

A function to output a ANSI code to a formatter to set the background to this color

fn fmt_raw_ansi_fg(&self, f: &mut Formatter<'_>) -> Result

A function to output a raw ANSI code to a formatter to set the foreground to this color, but without including the ANSI delimiters.

fn fmt_raw_ansi_bg(&self, f: &mut Formatter<'_>) -> Result

A function to output a raw ANSI code to a formatter to set the background to this color, but without including the ANSI delimiters.